debian.aws: re-enable AUFS filesystem in kernel
AUFS became disabled in Ubuntu 22.04, but SPS uses it for merging a
writable layer onto the firmware on squashfs, and we would rather not
switch to overlayfs at this point yet.

With this change the AUFS feature can be turned on with 'aufs' kernel
module as before.
This patch is based on the config of Ubuntu 20.04 (focal).
In this package the kernel config is generated from the annotations
files so in this patch do not contains explicit kernel config
modifications. If there is no `config.common.ubuntu` config file, then the
build will generate the config with the `debian/scripts/misc/annotations`
script.
You can learn more from the `debian/rules.d/2-binary-arch.mk`.

The new annotations model description can be found here:
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/2019000

References: azure #351715